Update SearchComponent
Refractor autocomplete event to SearchComponent, because it's a better place for it to live.
