+++ src/refinement/refinement.component.handlebars
@@ -1,5 +1,5 @@
 <h3>Merk</h3>
-<ul>
+<ul class="list-brands">
     {{#each merken}}
     <li>
         <input type="checkbox" name="chkBoxRefinement" value="{{value}}" /> {{label}}
