+++ src/refinement/style.scss
@@ -1,5 +1,5 @@
 @import "../_variables.scss";  
-ul{
+ul.list-brands{
     list-style: none;
     padding: 0;
 }
