+++ src/search/search.component.js
@@ -1,4 +1,5 @@
 import $ from "jquery";
+import "jquery-ui/ui/widgets/autocomplete";
 import "./style.scss";
 
 let template = require("./search.component.handlebars");
@@ -15,7 +16,37 @@ class SearchComponent {
     constructor(container) {
         this.container = container;
 
+        this.bindAutocompleteToDomContainer();
+
         this.bindHandlebarsTemplateToDom();
+    }
+
+    /**
+     * Bind autocomplete event to the search input field to enable search suggestions
+     */
+    bindAutocompleteToDomContainer() {
+        // When the DOM is ready...
+        $(() => {
+            // Bind autocomple event to the searchbox
+            // This enables the application to generate search suggestions
+            $("#searchfield").autocomplete({
+                source: (request, response) => {
+                    $.ajax({
+                        url: `${window.location.protocol}//zoeksuggesties.s-bol.com/extern/qs/OpenSearchJSCB/search_suggestions_callback/media_all/${encodeURIComponent(request.term)}`,
+                        cache: true,
+                        dataType: "jsonp",
+                        jsonp: false,
+                        jsonpCallback: "search_suggestions_callback",
+                        success: (data) => {
+                            let suggestions = [];
+                            data[1].forEach(suggestion => { suggestions.push(suggestion); });
+                            response(suggestions);
+                        },
+                        error: (error) => response([])
+                    });
+                }
+            });
+        });
     }
 
     /**
@@ -23,8 +54,7 @@ class SearchComponent {
      */
     bindHandlebarsTemplateToDom() {
         $(this.container).html(template());
-
-
     }
+
 }
 export default SearchComponent;
